pyproject.toml
[project]
name = "ai-novel-factory"
version = "1.0.0"
description = "AI-powered novel writing factory with multi-agent system"
dependencies = [
    "streamlit>=1.31.0",
    "crewai>=0.75.0",
    "litellm>=1.40.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.1.0",
    "ebooklib>=0.18",
    "pypandoc>=1.13",
    "pyyaml>=6.0.1",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
